le-services: Use syslog
- Avoid using log.h, instead use syslog.h
- Clean up compilation warnings related to logs
- Remove logs from interface headers to plugin

+++ recorder/src/service/qmmf_camera_source.cc
@@ -369,7 +369,8 @@ bool CameraSource::ValidateSlaveTrackParam(
   QMMF_DEBUG("%s %d x %d -> %d x %d fmt 0x%x -> 0x%x", __func__,
       master_params.width, master_params.height,
       slave_params.width, slave_params.height,
-      master_params.format, slave_params.format);
+      static_cast<int32_t>(master_params.format),
+      static_cast<int32_t>(slave_params.format));
 
   if ((slave_params.format != VideoFormat::kNV12) &&
       (slave_params.format != VideoFormat::kNV12UBWC) &&
@@ -398,7 +399,8 @@ bool CameraSource::CheckLinkedStream(
   QMMF_DEBUG("%s %d x %d -> %d x %d fmt 0x%x -> 0x%x", __func__,
     master_params.width, master_params.height,
     slave_params.width, slave_params.height,
-    master_params.format, slave_params.format);
+    static_cast<int32_t>(master_params.format),
+    static_cast<int32_t>(slave_params.format));
 
   if ((slave_params.format != VideoFormat::kNV12) &&
       (slave_params.format != VideoFormat::kNV12UBWC) &&
@@ -815,7 +817,7 @@ const shared_ptr<TrackSource>& CameraSource::GetTrackSource(uint32_t track_id) {
 
 bool CameraSource::IsTrackIdValid(const uint32_t track_id) {
 
-  QMMF_DEBUG("%s: Number of Tracks exist: %d",__func__, track_sources_.size());
+  QMMF_DEBUG("%s: Number of Tracks exist: %ld",__func__, track_sources_.size());
   return (track_sources_.count(track_id) != 0) ? true : false;
 }
 
@@ -1363,7 +1365,7 @@ status_t TrackSource::StopTrack(bool cached) {
   QMMF_INFO("%s: Pipe stop done(%x)", __func__, id_);
   {
     std::lock_guard<std::mutex> lk(buffer_list_lock_);
-    QMMF_DEBUG("%s: Track(%x): buffer_list_.size(%d)", __func__,
+    QMMF_DEBUG("%s: Track(%x): buffer_list_.size(%ld)", __func__,
         id_, buffer_list_.size());
   }
 
@@ -1410,7 +1412,7 @@ void TrackSource::OnFrameAvailable(StreamBuffer& buffer) {
   }
 
   if (IsStop()) {
-    QMMF_DEBUG("%s: Track(%x): Stop is triggred, return buffer fd: %d ts: %lld",
+    QMMF_DEBUG("%s: Track(%x): Stop is triggred, return buffer fd: %d ts: %ld",
         __func__, id_, buffer.fd, buffer.timestamp);
 
     std::unique_lock<std::mutex> lock(frame_lock_);
@@ -1477,7 +1479,7 @@ status_t TrackSource::ReturnTrackBuffer(std::vector<BnBuffer>& bn_buffers) {
 
   std::unique_lock<std::mutex> lock(frame_lock_);
   for (size_t i = 0; i < bn_buffers.size(); ++i) {
-    QMMF_VERBOSE("%s: Track(%x): bn_buffers[%d].ion_fd=%d", __func__,
+    QMMF_VERBOSE("%s: Track(%x): bn_buffers[%ld].ion_fd=%d", __func__,
         id_, i, bn_buffers[i].ion_fd);
 
     std::lock_guard<std::mutex> autoLock(buffer_list_lock_);
@@ -1489,7 +1491,7 @@ status_t TrackSource::ReturnTrackBuffer(std::vector<BnBuffer>& bn_buffers) {
     ReturnBufferToProducer(buffer);
     buffer_list_.erase(it);
   }
-  QMMF_DEBUG("%s: Track(%x): buffer count still with client = %d", __func__,
+  QMMF_DEBUG("%s: Track(%x): buffer count still with client = %ld", __func__,
       id_, buffer_list_.size());
 
   if (buffer_list_.size() == 0) {
@@ -1535,11 +1537,11 @@ void TrackSource::EnableFrameRepeat(const bool enable) {
 }
 
 void TrackSource::ReturnBufferToProducer(StreamBuffer& buffer) {
-  QMMF_DEBUG("%s: Enter Track(%x): fd: %d ts: %lld", __func__,
+  QMMF_DEBUG("%s: Enter Track(%x): fd: %d ts: %ld", __func__,
       id_, buffer.fd, buffer.timestamp);
 
   if (buffer_map_.find(buffer.handle) == buffer_map_.end()) {
-    QMMF_ERROR("%s: Track(%x): fd: %d ts: %lld", __func__,
+    QMMF_ERROR("%s: Track(%x): fd: %d ts: %ld", __func__,
         id_, buffer.fd, buffer.timestamp);
   } else {
     QMMF_DEBUG("%s: Buffer is back to Producer Intf,buffer(0x%p) RefCount=%d",
@@ -1561,7 +1563,7 @@ void TrackSource::ReturnBufferToProducer(StreamBuffer& buffer) {
 }
 
 void TrackSource::NotifyBufferReturned(StreamBuffer& buffer) {
-  QMMF_DEBUG("%s: Enter Track(%x): fd: %d ts: %lld", __func__,
+  QMMF_DEBUG("%s: Enter Track(%x): fd: %d ts: %ld", __func__,
       id_, buffer.fd, buffer.timestamp);
   std::unique_lock<std::mutex> lock(frame_lock_);
   ReturnBufferToProducer(buffer);
